refactor: extracted logic and created hoooks, then imported, update import link
